+++ datos/borrar_institucion.php
@@ -1,7 +1,14 @@
 <?php
-    include('conex.inc');
+if(isset($_POST['cod']))
+{
+	include('conex.php');
     $id_institucion 	= $_POST['cod'];
     $sql = "DELETE FROM instituciones WHERE id_institucion=".$id_institucion;
 
     $eliminar = mysqli_query($db,$sql);
+}
+else
+{
+	header("location: error.php");
+}
 ?>
